% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsmsampling.R
\name{m_ran_sampling}
\alias{m_ran_sampling}
\title{RSM Orb Monetary size Random Sampling}
\usage{
m_ran_sampling(pop, am, SR, RC, PL, PM, EA)
}
\arguments{
\item{pop}{Population Dataset.}

\item{am}{Amount Column Name : Character.}

\item{SR}{Significant Risk : "Yes" or "No".}

\item{RC}{Reliance on Control : "Yes" or "No".}

\item{PL}{Planned Level of Assurance from Substantive Analytical Procedures : "Analytical.Procedures.Not.Performed"-4 , "High"-1 , "Low"-2, "Moderate"-3.}

\item{PM}{Tolerable Misstatement(generally performance materiality) : Integer number.}

\item{EA}{Expected misstatement : generally 0.05.}
}
\value{
Sampling result data.frame.
}
\description{
RSM Orb Monetary size Random Sampling
}
\examples{
m_ran_sampling(pop, "CR", "Yes", "Yes", 4, 700000000, 0.05)
}
