% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop}
\alias{pop}
\title{Example of Revenue Journal Entry data}
\format{
## `pop`
A data frame with 57,419 rows and 22 columns:
\describe{
  \item{전표일자}{Journal Entry date}
  \item{전표번호}{Journal Entry number}
  \item{차변금액}{Debit Record}
  \item{대변금액}{Credit Record}
  ...
}
}
\source{
<https://www.kicpa.or.kr/>
}
\usage{
pop
}
\description{
A subset of data from the KICPA Cyber training center
}
\keyword{datasets}
