% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsmsampling.R
\name{ran_sampling}
\alias{ran_sampling}
\title{Random Sampling by input Sample Size}
\usage{
ran_sampling(pop, n)
}
\arguments{
\item{pop}{Population Dataset.}

\item{n}{Sample Size.}
}
\value{
Sampling result data.frame.
}
\description{
Random Sampling by input Sample Size
}
\examples{
ran_sampling(pop, 25)
}
